test/udp_rendezvous_server: add integration test
Also did some code modifications to make code testable:
1. exposed UdpRendezvousClient,
2. implemented UdpRendezvousServer constructor with given UdpSock so that we could know what port will be used.
